+++ sdk/core/src/shared/getOptionsFromEnvironment.ts
@@ -46,6 +46,8 @@ type EnvironmentOptions = {
   phoneIssuerBasePath?: string
   emailIssuerBasePath?: string
   revocationUrl?: string
+  userPoolId?: string
+  clientId?: string
 }
 
 function getBasicOptionsFromEnvironment(options: EnvironmentOptions) {
@@ -68,8 +70,8 @@ function getBasicOptionsFromEnvironment(options: EnvironmentOptions) {
     case 'dev':
       return {
         env: env as Env,
-        clientId: DEV_COGNITO_CLIENT_ID,
-        userPoolId: DEV_COGNITO_USER_POOL_ID,
+        clientId: options.clientId || DEV_COGNITO_CLIENT_ID,
+        userPoolId: options.userPoolId || DEV_COGNITO_USER_POOL_ID,
         phoneIssuerBasePath: options.phoneIssuerBasePath || DEV_PHONE_ISSUER_BASE_PATH,
         emailIssuerBasePath: options.emailIssuerBasePath || DEV_EMAIL_ISSUER_BASE_PATH,
         ...urls,
@@ -79,8 +81,8 @@ function getBasicOptionsFromEnvironment(options: EnvironmentOptions) {
     case 'prod':
       return {
         env: env as Env,
-        clientId: PROD_COGNITO_CLIENT_ID,
-        userPoolId: PROD_COGNITO_USER_POOL_ID,
+        clientId: options.clientId || PROD_COGNITO_CLIENT_ID,
+        userPoolId: options.userPoolId || PROD_COGNITO_USER_POOL_ID,
         phoneIssuerBasePath: options.phoneIssuerBasePath || PROD_PHONE_ISSUER_BASE_PATH,
         emailIssuerBasePath: options.emailIssuerBasePath || PROD_EMAIL_ISSUER_BASE_PATH,
         ...urls,
@@ -89,8 +91,8 @@ function getBasicOptionsFromEnvironment(options: EnvironmentOptions) {
     case 'staging':
       return {
         env: env as Env,
-        clientId: STAGING_COGNITO_CLIENT_ID,
-        userPoolId: STAGING_COGNITO_USER_POOL_ID,
+        clientId: options.clientId || STAGING_COGNITO_CLIENT_ID,
+        userPoolId: options.userPoolId || STAGING_COGNITO_USER_POOL_ID,
         phoneIssuerBasePath: options.phoneIssuerBasePath || STAGING_PHONE_ISSUER_BASE_PATH,
         emailIssuerBasePath: options.emailIssuerBasePath || STAGING_EMAIL_ISSUER_BASE_PATH,
         ...urls,
@@ -113,6 +115,8 @@ const splitOptions = <TOptions extends SdkOptions>(options: TOptions) => {
     emailIssuerBasePath,
     revocationUrl,
     storageRegion,
+    clientId,
+    userPoolId,
     ...otherOptions
   } = options
 
@@ -132,6 +136,8 @@ const splitOptions = <TOptions extends SdkOptions>(options: TOptions) => {
       phoneIssuerBasePath,
       emailIssuerBasePath,
       revocationUrl,
+      clientId,
+      userPoolId,
     },
     storageRegion,
     otherOptions,