+++ sdk/core/test/unit/shared/index.ts
@@ -1,3 +1,4 @@
 require('./SdkErrorFromCode.test')
 require('./ParametersValidator.test')
 require('./sessionStorageHandler.test')
+require('./getOptionsFromEnvironment.test')